requirements.txt
eval_type_backport>=0.1.3
fastapi==0.111.0
uvicorn[standard]==0.29.0
sqlalchemy==2.0.30
psycopg[binary]>=3.1
alembic==1.13.1
pydantic-settings==2.2.1
pyjwt==2.8.0
twilio==9.0.4
sendgrid==6.11.0
pytest==8.2.0
pytest-cov==5.0.0
httpx==0.27.0
anthropic>=0.25.0
